offst-connection: Refactoring.
