+++ components/connection/src/lib.rs
@@ -9,8 +9,8 @@
 
 extern crate log;
 
-mod encrypt_keepalive;
-mod secure_connector;
+mod transforms;
 
-pub use self::encrypt_keepalive::create_encrypt_keepalive;
-pub use self::secure_connector::{create_secure_connector, create_version_encrypt_keepalive};
+pub use self::transforms::{
+    create_encrypt_keepalive, create_secure_connector, create_version_encrypt_keepalive,
+};
